nharker-server: separate route into different files
